// backend/prisma/seed_data.d.ts
export declare const MASTER_COLLEGES: {
    name: string;
    city: string;
    state: string;
    ownership: string;
    established: number;
    fees: number;
    rating: number;
    placementPercent: number;
    avgPackage: number;
    exams: string[];
    degrees: string[];
    description: string;
    slug: string;
    website: string;
}[];
//# sourceMappingURL=seed_data.d.ts.map